% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_cl.R
\name{CalEvi2Point}
\alias{CalEvi2Point}
\title{Calculate EVI2 time series from the downloaded point time series file.}
\usage{
CalEvi2Point(ard_pt_csv, out_file)
}
\arguments{
\item{ard_pt_csv}{The Landsat ARD point time series data in csv format.}

\item{out_file}{The output file path and name.}
}
\value{

}
\description{
Calculate EVI2 time series from the downloaded point time series file.
}
\examples{
\dontrun{

}
}
