% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{CheckTaskStatus}
\alias{CheckTaskStatus}
\title{Check the status of current tasks}
\usage{
CheckTaskStatus(token, limit, task_name = NULL, brief = FALSE)
}
\arguments{
\item{token}{AppEEARS login token. Use \code{Login()} to get a token.}

\item{limit}{The query limit amount.}

\item{task_name}{The name of the task.}

\item{brief}{Logical. \code{TRUE} to show brief information, \code{FALSE} to show the full
information (default).}
}
\value{
Job status information
}
\description{
Check the status of current tasks
}
\examples{
\dontrun{

}
}
