% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{DownloadTask}
\alias{DownloadTask}
\title{Download the job from AppEEARS.}
\usage{
DownloadTask(token, task_id, out_dir)
}
\arguments{
\item{token}{AppEEARS login token. Use \code{Login()} to get a token.}

\item{task_id}{The ID of the task. Use \code{CheckTaskStatus()} to get the IDs.}

\item{out_dir}{The target directory to store downloaded files.}
}
\value{

}
\description{
Download the job from AppEEARS.
}
\examples{
\dontrun{

}
}
