% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{FitAvgModel}
\alias{FitAvgModel}
\title{Fit the averaged model and get the model parameters.}
\usage{
FitAvgModel(date_vec, vi_vec, ifplot = FALSE)
}
\arguments{
\item{date_vec}{the date vector, be sure to convert the vector to "Date"
format or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}

\item{ifplot}{logical. Plot the model fit if TRUE.}
}
\description{
Fit the averaged model and get the model parameters.
}
