% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{FitBLSP}
\alias{FitBLSP}
\title{Bayesian mixed hierarchical land surface phenology model.}
\usage{
FitBLSP(
  date_vec,
  vi_vec,
  weights_vec = NULL,
  initValues = NULL,
  ifplot = FALSE
)
}
\arguments{
\item{date_vec}{The date vector, be sure to convert the vector to "Date" format
or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}

\item{weights_vec}{For specifying weights to observations. For example,
lower weights can be set to observations with snow.}

\item{initValues}{Initial values for MCMC sampling. We get these values from
fitting the averaged model. It could also be \code{NULL}.}

\item{ifplot:}{logical. Plot the model fit if TRUE. Note that the fitted curve
with CI will only be returned when \code{ifplot} is TRUE.}
}
\value{
A list of 2 data tables. \code{fitted} is the , while \code{phenos} contains the
estimated DOY of midgreenup and midgreendown per year, with upper and lower
confidence intervals. Note that \code{fitted} returns \code{NULL} if \code{ifplot} is FALSE.
}
\description{
Bayesian mixed hierarchical land surface phenology model.
}
