% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{FormatAvgData}
\alias{FormatAvgData}
\title{Format input date and VI vectors to the structure needed for fitting averaged
phenology models such as Fisher et al 2006, Elmore et al 2012.}
\usage{
FormatAvgData(date_vec, vi_vec)
}
\arguments{
\item{date_vec}{the date vector, be sure to convert the vector to "Date"
format or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}
}
\value{
A list that contains formated data.
}
\description{
Format input date and VI vectors to the structure needed for fitting averaged
phenology models such as Fisher et al 2006, Elmore et al 2012.
}
