% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{GetModel}
\alias{GetModel}
\title{Get the user selected model structure and parameters}
\usage{
GetModel(model_name)
}
\arguments{
\item{model_name}{The name of the model. For now only support \code{dblog7} and
\code{dblog6} corresponding to the 7- and 6-parameter double-logistic functions.}
}
\value{
A list containing model structure and parameters.
}
\description{
Get the user selected model structure and parameters
}
