% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{GetPhenosIdx}
\alias{GetPhenosIdx}
\title{Generate pheno from the predicted curve.
Only supports Elmore model (The double-logistic model used in BLSP). This function is used inside the Fit_BLSP function.}
\usage{
GetPhenosIdx(equation, params, t)
}
\arguments{
\item{equation}{The model equation.}

\item{params}{The Parameter list.}

\item{t}{Date vector.}
}
\value{
The phenological timing in day of year (DOY)
}
\description{
Generate pheno from the predicted curve.
Only supports Elmore model (The double-logistic model used in BLSP). This function is used inside the Fit_BLSP function.
}
