% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsatEVI2}
\alias{landsatEVI2}
\title{Landsat EVI2 time series.}
\format{
A data frame with 802 rows and 3 variables:
\describe{
\item{date}{Observation dates}
\item{evi2}{The two-band enhanced vegetation index value}
\item{snow}{Logical, indicates whether the observation contains snow}
}
}
\usage{
landsatEVI2
}
\description{
An Example EVI2 time series from Landsat observations.
}
\keyword{datasets}
