% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{Login}
\alias{Login}
\title{Log in to AppEEARS system.}
\usage{
Login(usr, pwd)
}
\arguments{
\item{usr, pwd}{Username and password of AppEEARS system.}
}
\value{
Return the token string
}
\description{
Log in to AppEEARS system.
}
