% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{Logout}
\alias{Logout}
\title{log out from AppEEARS.}
\usage{
Logout(token)
}
\arguments{
\item{token}{AppEEARS login token. Use \code{Login()} to get a token.}
}
\value{

}
\description{
log out from AppEEARS.
}
