% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fit.R
\name{PlotBLSP}
\alias{PlotBLSP}
\title{Plot BLSP model fitting result.}
\usage{
PlotBLSP(blsp_fit, if_return_fit = FALSE)
}
\arguments{
\item{blsp_fit}{The object of \code{BlspFit} class returned by \code{FitBLSP()} function.}

\item{if_return_fit}{Logic. Determine whether return the fitted values. Default
is \code{FALSE}.}
}
\value{
A plot showing the BLSP model fitting result. If \code{if_return_fit} is true,
the model fitted time series as well as the 95\% credible interval will also be
returned.
}
\description{
Plot BLSP model fitting result.
}
\examples{
\dontrun{
blsp_fit <- FitBLSP(landsatEVI2$date, landsatEVI2$evi2, verbose = TRUE)
fitted_dt <- PlotBLSP(blsp_fit, if_return_fit = TRUE)
}
}
