% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{ReqArdPoint}
\alias{ReqArdPoint}
\title{Request Landsat ARD point time series data using AppEEARS.}
\usage{
ReqArdPoint(
  pts_df,
  task_name,
  token,
  start_date = "01-01-1984",
  end_date = "12-31-2022"
)
}
\arguments{
\item{pts_df}{A data frame containing point information, including 4 columns:
\code{id}, \code{longitude}, \code{latitude}, \code{category}.}

\item{task_name}{Task name.}

\item{token}{AppEEARS login token. Use \code{Login()} to get a token.}
}
\value{
TRUE if the job is successfully submited.
}
\description{
Request Landsat ARD point time series data using AppEEARS.
}
\examples{
\dontrun{
lat <- c(36.206228, 37.289327) # Latitude of the point sites
lon <- c(-112.127134, -112.973760) # Longitude of the point sites
id <- c("0", "1") # ID for the point sites
pts_df <- data.frame(id = id, longitude = lon, latitude = lat, 
    category = category)
}
}
