% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlp_funs.R
\name{SanityCheck}
\alias{SanityCheck}
\title{Check if user inputs are reasonable}
\usage{
SanityCheck(date_vec, vi_vec, weights_vec, model, cred_int_level, init_values)
}
\arguments{
\item{date_vec}{The date vector, be sure to convert the vector to "Date"
format or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}

\item{weights_vec}{A numeric vector of same length as vi_vec specifying the
weights for the supplied observations. Must be between 0 and 1, inclusive.}

\item{model}{The model string.}

\item{cred_int_level}{A scalar value from 0 to 1 (exclusive) that specifies
the level for equal-tailed credible intervals of the estimated phenometrics.
The default level is 0.9, generating \verb{90\%} credible intervals. The end
points of these intervals define the upper and lower bounds for the estimated
phenometrics.}

\item{init_values}{Initial values for MCMC sampling. By default, it is
assgined \code{NULL}. It could also be an object returned from the \code{FitAvgModel()}
function that fits an averaged model or a numeric vector provided by the
user.}
}
\description{
Check if user inputs are reasonable
}
