% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_dl.R
\name{SubmitTask}
\alias{SubmitTask}
\title{Submit a point task. \code{task_type} can either be 'point' or 'area', but relative
params need to be provided.}
\usage{
SubmitTask(
  token,
  task_name,
  task_type = "point",
  start_date = NULL,
  end_date = NULL,
  recursive = FALSE,
  layers = NULL,
  point_df = NULL,
  polygon_file = NULL,
  out_format = "geotiff",
  out_proj = NULL
)
}
\arguments{
\item{token}{AppEEARS login token. Use \code{Login()} to get a token.}

\item{task_name}{The name of the task.}

\item{task_type}{The AppEEARS task type, either "point" or "area".}

\item{start_date, end_date}{The start and end dates of the requested time series.}

\item{recursive}{If the specified start and end dates are recursive. See AppEEARS
documentation.}

\item{layers}{The interested layers.}

\item{out_format}{The desired image format, default is \code{geotiff}.}

\item{out_proj}{The desired image projection.}
}
\value{

}
\description{
Submit a point task. \code{task_type} can either be 'point' or 'area', but relative
params need to be provided.
}
\examples{
\dontrun{

}
}
