% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Transparent}
\alias{Transparent}
\title{Make a standard color transparent.
This function is borrowed from 'yarrr' package, but I changed the trans.val
to use alpha value directly.}
\usage{
Transparent(orig.col, alpha = 1, maxColorValue = 255)
}
\arguments{
\item{orig.col:}{the original color, can be a color name, a hexadecimal code,
or a rgb vector.}

\item{alpha:}{define the transparent level.}

\item{maxColorValue:}{used to convert the color to rgb format before making
it transparent.}
}
\value{
color code.
}
\description{
Make a standard color transparent.
This function is borrowed from 'yarrr' package, but I changed the trans.val
to use alpha value directly.
}
