% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{FitAvgModel}
\alias{FitAvgModel}
\title{Fit the averaged model and get the model parameters.}
\usage{
FitAvgModel(date_vec, vi_vec)
}
\arguments{
\item{date_vec}{the date vector, be sure to convert the vector to "Date"
format or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}
}
\value{
Model parameters to be used as MCMC initial parameters in
the FitBLSP function.
}
\description{
Fit the averaged model and get the model parameters.
}
