% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{FitBLSP}
\alias{FitBLSP}
\title{Fit a Bayesian mixed hierarchical land surface phenology model.}
\usage{
FitBLSP(
  date_vec,
  vi_vec,
  weights_vec = NULL,
  initValues = NULL,
  verbose = FALSE
)
}
\arguments{
\item{date_vec}{The date vector, be sure to convert the vector to "Date"
format or use "yyyy-mm-dd" format string.}

\item{vi_vec}{The vegetation index vector.}

\item{weights_vec}{A numeric vector of same length as vi_vec specifying the
weights for the supplied observations. Must be between 0 and 1, inclusive.}

\item{initValues}{Initial values for MCMC sampling. By default, it is
assgined \code{NULL}. It could also be an object returned from the \code{FitAvgModel()}
function that fits an averaged model or a numeric vector provided by the user.}

\item{verbose}{logical. If \code{TRUE}, the progress will be reported.}
}
\value{
An object of class \code{BlspFit} will be returned. The object contains the
estimated spring and autumn phenometrics for each year, the generated model
parameter samples, and the input data.
}
\description{
This function fits a Bayesian mixed hierarchical land surface phenology model
to the supplied data (can be sparse), and returns phenometrics for the
entire time frame. For further explanation, please see the vignette.
}
\examples{
\dontrun{
data(landsatEVI2)
blsp_fit <- FitBLSP(date_vec = landsatEVI2$date, vi_vec = landsatEVI2$evi2)
}
}
